Add custom rois and offset when loading hdf5 files
Currently, offsets for volume rois are always set to 0, 0, 0. This commits allows the user to specify custom rois or if not specified, rois are inferred from the datasource. As a second feature, the user can also specify a dataset specific offset that tells how the different volumes are aligned relative to each other.